refactor(ui): extract AppRouter and slide_style into app::router/animations
