+++ src/ui/app/mod.rs
@@ -1,9 +1,12 @@
+mod animations;
 mod boot;
 mod contexts;
+mod router;
 mod watchers;
 
 pub use boot::{load_consent, load_lang};
 pub use contexts::{use_app_contexts, AppContexts};
+pub use router::AppRouter;
 pub use watchers::{
     use_history_tracker, use_picker_reset_on_view, use_sync_watcher,
     use_transcription_watcher,
